+++ CHANGELOG.md
@@ -4,7 +4,20 @@
 ### Breaking
 
 ### Features
+
 ### Improvements
+* implement abstract Event class to encapsulate event data, processing state, warnings, and errors
+* integrate dotted field handling methods directly into Event, enabling structured field access and manipulation
+* support event identity and hashability based on data, allowing usage in sets and as dictionary keys
+
+* add ng packages as namespace in dirs 'unit' and 'logprep' as preparation for new architecture implementation
+* add abstract EventMetaclass
+* implement KafkaInputMetadata
+
+* implement EventState class to manage the lifecycle of log events
+* integrate a finite state machine to control valid state transitions
+* support for both automatic and conditional (success-based) state transitions
+
 ### Bugfix
 
 ## 17.0.0